requirements.txt
# Image Processing
numpy
pillow
scikit-learn

# Interacting with API
ping3
aiohttp
requests

# Database
prisma

# Discord and Environment Management
discord.py
psutil
python-dotenv
setuptools